+++ src/marshmallow_sqlalchemy/convert.py
@@ -458,7 +458,10 @@ class ModelConverter:
         nullable = True
         for pair in prop.local_remote_pairs:
             if not pair[0].nullable:
-                if prop.uselist is True:
+                if (
+                    prop.uselist is True
+                    or self.DIRECTION_MAPPING[prop.direction.name] is False
+                ):
                     nullable = False
                 break
         kwargs.update({"allow_none": nullable, "required": not nullable})